contracts: economic simulation: Get hints from contracts in advance
Still keeping the local order to use it as approx hints. But calling
`findInsertPosition` before as a static call, to avoid the trove
operation reverting if the hints were not correct.
The local order is only based on target CR, but not update with trove
adjustments, so it’s approximate.